Do not register the B1 map to other structural scans
The B1 map is valid only with reference to the unmodified scanner coordinates,
so registering it to another scan is not a good thing. This registration would
give results very close to identity anyway, so it is not expected to have
caused serious issues with specimens that were processed with the old script.